wifi: use uint16_t to count MPDUs in A-MPDU
Since IEEE 802.11ax standard increased maximum Block Ack Bitmap length
to 256, and HeConfiguration::MpduBufferSize can be set to value 256,
numbers of MPDUs in block ack manager and rate control should be
stored in uint16_t.

Before this patch, when 256 MPDUs were received in one A-MPDU,
nSuccessfulMpdus counter overflowed and was accounted as 0.

+++ src/wifi/model/rate-control/ideal-wifi-manager.h
@@ -69,7 +69,7 @@ private:
                       double ctsSnr, WifiMode ctsMode, double rtsSnr);
   void DoReportDataOk (WifiRemoteStation *station, double ackSnr, WifiMode ackMode,
                        double dataSnr, uint16_t dataChannelWidth, uint8_t dataNss);
-  void DoReportAmpduTxStatus (WifiRemoteStation *station, uint8_t nSuccessfulMpdus, uint8_t nFailedMpdus,
+  void DoReportAmpduTxStatus (WifiRemoteStation *station, uint16_t nSuccessfulMpdus, uint16_t nFailedMpdus,
                               double rxSnr, double dataSnr, uint16_t dataChannelWidth, uint8_t dataNss);
   void DoReportFinalRtsFailed (WifiRemoteStation *station);
   void DoReportFinalDataFailed (WifiRemoteStation *station);
